requirements-data.txt
# Data pipeline for dashboard JSON generation
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
# Optional: higher-frequency BTC via ccxt
# ccxt>=4.0.0
